rating werkt en getest, 1 dto voor gegevens gebruikt, Artwork gerelateerde requests ook verplaatst naar RatingController.
