+++ src/main/java/nl/novi/theartroom/controllers/RatingController.java
@@ -1,5 +1,6 @@
 package nl.novi.theartroom.controllers;
 
+import nl.novi.theartroom.dtos.RatingAverageOutputDto;
 import nl.novi.theartroom.dtos.RatingDto;
 import nl.novi.theartroom.models.Rating;
 import nl.novi.theartroom.services.RatingService;
@@ -50,5 +51,26 @@ public class RatingController {
     public ResponseEntity<Void> deleteRating(@PathVariable Long ratingId) {
         ratingService.deleteRating(ratingId);
         return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/{artworkId}/ratings")
+    public ResponseEntity<List<RatingDto>> getRatingsForArtwork(@PathVariable Long artworkId) {
+        List<RatingDto> ratings = ratingService.getRatingsForArtwork(artworkId);
+
+        return ResponseEntity.ok(ratings);
+    }
+
+    @PostMapping("/{artworkId}/ratings")
+    public ResponseEntity<Void> addRatingToArtwork(@PathVariable Long artworkId, @RequestBody RatingDto rating) {
+        ratingService.addRatingToArtwork(artworkId, rating.getRating(), rating.getComment());
+        return ResponseEntity.created(null).build();
+    }
+
+    // Get an average rating for an artwork
+    @GetMapping("/{artworkId}/ratings/average")
+    public ResponseEntity<RatingAverageOutputDto> getAverageRatingForArtwork(@PathVariable Long artworkId) {
+        double averageRating = ratingService.calculateAverageRatingForArtwork(artworkId);
+        RatingAverageOutputDto outputDto = new RatingAverageOutputDto(averageRating);
+        return ResponseEntity.ok(outputDto);
     }
 }
